Add support for quilt if available.
The patches will be applied using quilt, which makes modifying existing
patches and adding new patches a lot easier.  From the point of JDK
development, mercurial queues might have been the more appropriate solution.
But we are trying to keep things simple, as our main goal is making JDK
building possible for a large audience.  The quilt patches live in the git
repository just as they have before, and will be applied using patch just as
before unless the user actually has quilt installed.  Maintaining them will
be a bit easier now, though.

In the long run it might make sense to support different series files, so
that e.g. for a given tag name a suitable set of patches can be configured
which is known to work for that tag.  But I guess we'll deal with that once
the need actually arises.