Avoid using Ruby.
There are tow ruby scripts used in building JObjC. They are however
little more than shell scripts, and the Ruby dependency is quite
strict as to which versions do work, so it has already caused trouble
in the past. This patch here gets rid of Ruby as a dependency
altogether, using bash instead.